修改一下内容: 1.子单优惠分拆调整,添加了shareDiscount计算 2.调整发货单全部发货之后才通知电商收货,电商确认收货之后自动通知恒康发货单收货时间 3.调整同步恒康时分转为元金额转换过程中小数被忽略的问题 4.子单存在部分退款,此时需要其他待处理的子单自动生成发货单 5.售后单完善之后自动同步恒康 6.添加发货单确认收货之后同步恒康失败之后手动触发的api

+++ middle-web/src/main/java/com/pousheng/middle/hksyc/dto/trade/SycHkRefund.java
@@ -3,6 +3,7 @@ package com.pousheng.middle.hksyc.dto.trade;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * Created by songrenfei on 2017/7/19
@@ -17,12 +18,12 @@ public class SycHkRefund implements Serializable{
     private String shopId;
     private String stockId;
     private String performanceShopId;
-    private Integer refundOrderAmount;
+    private BigDecimal refundOrderAmount;
     private Integer refundFreight;
     private String type;
     private String status;
     private String createdDate;
-    private Integer totalRefund;
+    private BigDecimal totalRefund;
     private String logisticsCompany;
     private String logisticsCode;
     private String memo;
